@@ -53,18 +53,23 @@
     "xmldom": "^0.5.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.17.0",
+    "@babel/eslint-parser": "^7.19.1",
     "@intlify/eslint-plugin-vue-i18n": "^1.4.1",
     "@nuxt/types": "^2.16.0",
-    "@nuxtjs/eslint-config": "^9.0.0",
+    "@nuxt/typescript-build": "^2.1.0",
+    "@nuxtjs/eslint-config-typescript": "^12.0.0",
     "@nuxtjs/eslint-module": "^3.1.0",
     "@nuxtjs/localforage": "^1.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.51.0",
+    "@typescript-eslint/parser": "^5.51.0",
+    "@vue/cli-plugin-typescript": "^5.0.8",
+    "@vue/eslint-config-typescript": "^11.0.2",
     "@vue/test-utils": "^1.3.0",
     "autoprefixer": "^10.4.13",
     "babel-jest": "^24.1.0",
-    "eslint": "^8.14.0",
+    "eslint": "^8.34.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-nuxt": "^3.2.0",
+    "eslint-plugin-nuxt": "^4.0.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-webpack-plugin": "^3.1.1",
     "fibers": "^5.0.0",
@@ -72,6 +77,7 @@
     "node-sass": "^6.0.1",
     "prettier": "^2.6.2",
     "sass-loader": "^10.2.1",
+    "typescript": "^4.9.5",
     "vue-server-renderer": "^2.6.13"
   }
 }
